added error status codes for invalid requests

+++ routes/index.js
@@ -11,13 +11,25 @@ router.get("/", async (req, res) => {
 // POST
 router.post("/", async (req, res) => {
   const { title, body } = req.body;
-  const post = await Post.create({ title, body });
-  res.send(post);
+  if (!title || !body)
+    // Bad request code
+    return res.status(400).send({ error: "Title or Body is missing..." });
+  try {
+    const post = await Post.create({ title, body });
+    // Created code
+    res.status(201).send(post);
+  } catch (e) {
+    // Server error code
+    res
+      .status(500)
+      .send({ error: "Title or Body needs a minimum of 5 characters" });
+  }
 });
 
 // GET ID
 router.get("/:id", async (req, res) => {
   const id = req.params.id;
+  if (!id) return res.status(400).send({ error: "Missing id..." });
   try {
     const post = await Post.findById(id);
     res.send(post);
@@ -30,21 +42,37 @@ router.get("/:id", async (req, res) => {
 
 router.put("/:id", async (req, res) => {
   const id = req.params.id;
+  if (!id) return res.status(400).send({ error: "Missing id..." });
+
   const { title, body } = req.body;
-  const post = await Post.findByIdAndUpdate(
-    id,
-    { $set: { title, body } },
-    { new: true }
-  );
-  res.send(post);
+
+  if (!title || !body)
+    return res.status(400).send({ error: "Title or Body is missing..." });
+
+  try {
+    const post = await Post.findByIdAndUpdate(
+      id,
+      { $set: { title, body } },
+      { new: true }
+    );
+    res.send(post);
+  } catch (e) {
+    res.status(400).send({ error: "Post was not modified" });
+  }
 });
 
 // DELETE ID
 
 router.delete("/:id", async (req, res) => {
   const id = req.params.id;
-  const post = await Post.findByIdAndDelete(id);
-  res.send(post);
+  if (!id) return res.status(400).send({ error: "Missing id..." });
+
+  try {
+    const post = await Post.findByIdAndDelete(id);
+    res.send(post);
+  } catch (e) {
+    res.status(400).send({ error: "Post was not deleted" });
+  }
 });
 
 module.exports = router;
